realm: Refactor rsi-test
