+++ realm/rsi-test/src/entry.rs
@@ -1,9 +1,3 @@
-const STACK_SIZE: usize = 0x1000;
-
-#[no_mangle]
-#[link_section = ".stack"]
-static mut STACK: [u8; STACK_SIZE] = [0; STACK_SIZE];
-
 #[link_section = ".head.text"]
 #[no_mangle]
 unsafe extern "C" fn _entry() -> ! {
