+++ scripts/build-kvm-unit-tests.sh
@@ -8,7 +8,6 @@ TOOLCHAIN="$ROOT/assets/toolchain"
 AARCH64="$TOOLCHAIN/aarch64-none-linux-gnu-10-2/bin/aarch64-none-linux-gnu-"
 
 cd $KVM_UNIT_TESTS
-#./configure --arch=arm64 --cross-prefix=$AARCH64 --target=kvmtool
+./configure --arch=arm64 --cross-prefix=$AARCH64 --target=kvmtool
 
 make
-cp -R arm/* $ROOT/out/shared/kvm-unit-tests
