+++ scripts/build-tf-rmm.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
+set -e
+
 ROOT=$(git rev-parse --show-toplevel)
 
 cd $ROOT/third-party/tf-rmm
-# git submodule update --init --recursive
 export CROSS_COMPILE=$1
 cmake -DRMM_CONFIG=fvp_defcfg -S . -B build
 cmake --build build
